+++ clients/fides-js/src/lib/init-utils.ts
@@ -111,6 +111,7 @@ export const getCoreFides = ({
       fidesConsentOverride: null,
       otFidesMapping: null,
       fidesDisabledNotices: null,
+      fidesDisabledSystems: null,
       fidesConsentNonApplicableFlagMode: null,
       fidesConsentFlagType: null,
       fidesInitializedEventMode: "once",